rename AplCheck file so its more consistent with other analyzers
